코드수정

+++ src/main/java/com/example/myproject/controller/MusicListController.java
@@ -3,11 +3,6 @@ package com.example.myproject.controller;
 import com.example.myproject.domain.*;
 import com.example.myproject.dto.*;
 import com.example.myproject.service.*;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jws;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
@@ -20,8 +15,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.util.WebUtils;
-import javax.crypto.SecretKey;
 import java.net.MalformedURLException;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
@@ -70,7 +63,7 @@ public class MusicListController {
                              @RequestParam(value = "page", defaultValue = "0") int page){
 
         String loginId = loginService.loginIdCheck(request, response);
-        log.info("로그인아이디static 확인 = {}", loginId);
+        log.info("로그인 아이디 확인 = {}", loginId);
         if (loginId != null){
             model.addAttribute("loginId", loginId);
         }
@@ -90,7 +83,7 @@ public class MusicListController {
                        @RequestParam(value = "page", defaultValue = "0") int page){
 
         String loginId = loginService.loginIdCheck(request, response);
-        log.info("로그인아이디static 확인 = {}", loginId);
+        log.info("로그인 아이디 확인 = {}", loginId);
         if (loginId != null){
             model.addAttribute("loginId", loginId);
         }
@@ -110,7 +103,7 @@ public class MusicListController {
                          HttpServletRequest request, Model model){
 
         String loginId = loginService.loginIdCheck(request, response);
-        log.info("로그인아이디static 확인 = {}", loginId);
+        log.info("로그인 아이디 확인 = {}", loginId);
         if (loginId != null){
             model.addAttribute("loginId", loginId);
         }
@@ -131,9 +124,12 @@ public class MusicListController {
         String referer = request.getHeader("Referer");
         String loginId = loginService.loginIdCheck(request, response);
         if (loginId == null){
+            log.info("댓글수정 실패 jwt 로그인 유지시간 초과");
             return "redirect:" + referer;
+        }else{
+            commentService.commentEdit(commentUpdateDto);
         }
-        commentService.commentEdit(commentUpdateDto);
+
 
         return "redirect:" + referer;
     }
@@ -143,10 +139,13 @@ public class MusicListController {
                                 HttpServletResponse response){
         String referer = request.getHeader("Referer");
         String loginId = loginService.loginIdCheck(request, response);
-        if (loginId == null){
+        if (loginId == null) {
+            log.info("댓글삭제 실패 jwt 로그인 유지시간 초과");
             return "redirect:" + referer;
+        }else{
+            commentService.commentDelete(commentDeleteDto);
         }
-        commentService.commentDelete(commentDeleteDto);
+
 
         return "redirect:" + referer;
     }
@@ -157,12 +156,13 @@ public class MusicListController {
         String referer = request.getHeader("Referer");
 
         String loginId = loginService.loginIdCheck(request, response);
+        log.info("로그인 아이디 확인 = {}", loginId);
         if (loginId == null){
+            log.info("대댓글등록 실패 jwt 로그인 유지시간 초과");
             return "redirect:" + referer;
+        }else{
+            commentService.replyAdd(commentReplyFormDto, loginId);
         }
-
-        log.info("로그인아이디static 확인 = {}", loginId);
-        commentService.replyAdd(commentReplyFormDto, loginId);
 
         return "redirect:" + referer;
     }
@@ -172,23 +172,24 @@ public class MusicListController {
         String referer = request.getHeader("Referer");
 
         String loginId = loginService.loginIdCheck(request, response);
+        log.info("로그인 아이디 확인 = {}", loginId);
         if (loginId == null){
+            log.info("댓글등록 실패 jwt 로그인 유지시간 초과");
             return "redirect:" + referer;
-        }
-        log.info("로그인아이디static 확인 = {}", loginId);
-
-        List<Comment> commentList = commentService.findCommentList(commentFormDto.getMusicListId());
-        if (commentList.isEmpty()){
-            int parent = 0;
-            Comment comment = commentService.commentAdd(commentFormDto, loginId, parent);
         }else{
-            List<Comment> byMusicListIdAndDivWidthSize =
-                    commentService.findByMusicListIdAndDivWidthSize(commentFormDto.getMusicListId());
-            int parent = byMusicListIdAndDivWidthSize.size();
-            Comment comment = commentService.commentAdd(commentFormDto, loginId, parent);
+            List<Comment> commentList = commentService.findCommentList(commentFormDto.getMusicListId());
+            if (commentList.isEmpty()){
+                int parent = 0;
+                Comment comment = commentService.commentAdd(commentFormDto, loginId, parent);
+            }else{
+                List<Comment> byMusicListIdAndDivWidthSize =
+                        commentService.findByMusicListIdAndDivWidthSize(commentFormDto.getMusicListId());
+                int parent = byMusicListIdAndDivWidthSize.size();
+                Comment comment = commentService.commentAdd(commentFormDto, loginId, parent);
+            }
         }
-
         return "redirect:" + referer;
+
     }
 
     @GetMapping("/content")
@@ -197,15 +198,9 @@ public class MusicListController {
             throws MalformedURLException {
 
         String loginId = loginService.loginIdCheck(request, response);
-        log.info("로그인아이디2확인해보자 = {}", loginId);
+        log.info("로그인 아이디 확인 = {}", loginId);
         if (loginId != null){
             model.addAttribute("loginId", loginId);
-        }
-        Cookie jwtCookie = WebUtils.getCookie(request, "jwtToken");
-        if (jwtCookie != null){
-            SecretKey key = Keys.hmacShaKeyFor("c3ByaW5nYm9vdC1qd3QtdHV0b3JpYWwtc3ByaW5nYm9vdC1qd3QtdHV0b3JpYWwtc3ByaW5nYm9vdC1qd3QtdHV0b3JpYWwK".getBytes(StandardCharsets.UTF_8));
-            Jws<Claims> claimsJws = Jwts.parser().verifyWith(key).build().parseSignedClaims(jwtCookie.getValue());
-            model.addAttribute("loginId",claimsJws.getPayload().getSubject());
         }
 
         LikeCount likeCount = likeCountService.findMyLike(id, loginId);
@@ -236,7 +231,7 @@ public class MusicListController {
         }
 
         if (fileList.isEmpty()) {
-            log.info("첨부파일없다.");
+            log.info("첨부파일 없음");
         }
 
         model.addAttribute("fileList", fileList);
@@ -246,8 +241,20 @@ public class MusicListController {
     }
 
     @PostMapping("/deleteMusicList")
-    public String deleteMusicList(MusicListDeleteDto musicListDeleteDto) {
-        musicListService.deleteMusicList(musicListDeleteDto.getMusicListId());
+    public String deleteMusicList(MusicListDeleteDto musicListDeleteDto, HttpServletRequest request,
+                                  HttpServletResponse response) {
+        String referer = request.getHeader("Referer");
+        String loginId = loginService.loginIdCheck(request, response);
+
+        log.info("로그인 아이디 확인 = {}", loginId);
+        loginId= null;
+        if (loginId == null){
+            log.info("뮤직리스트 삭제실패 jwt 로그인 유지시간 초과");
+            return "redirect:" + referer;
+        }else {
+            musicListService.deleteMusicList(musicListDeleteDto.getMusicListId());
+        }
+
         return "redirect:/";
     }
 
@@ -256,18 +263,22 @@ public class MusicListController {
                                           MusicListUpdateDto musicListUpdateDto, Model model, HttpServletResponse response) {
 
         String loginId = loginService.loginIdCheck(request, response);
-        log.info("로그인아이디static 확인 = {}", loginId);
-        if (loginId != null){
-            model.addAttribute("loginId", loginId);
-        }
-
-        model.addAttribute("musicListId", id);
-        List<FileList> fileList = fileListService.findByFiles(id);
-        Map<String, String> errors = musicListService.updateMusicList(id, musicListUpdateDto);
-        if (errors != null) {
-            model.addAttribute("fileList", fileList);
-            model.addAttribute("errors", errors);
-            return "musicList/editMusicListForm";
+        log.info("로그인 아이디 확인 = {}", loginId);
+        model.addAttribute("loginId", loginId);
+        if (loginId == null){
+            log.info("뮤직리스트 수정실패 jwt 로그인 유지시간 초과");
+        } else{
+            model.addAttribute("musicListId", id);
+            List<FileList> fileList = fileListService.findByFiles(id);
+            Map<String, String> errors = musicListService.updateMusicList(id, musicListUpdateDto);
+            if (errors != null) {
+                for (String value : errors.values()) {
+                    log.info("dto 검증 = {}", value);
+                }
+                model.addAttribute("fileList", fileList);
+                model.addAttribute("errors", errors);
+                return "musicList/editMusicListForm";
+            }
         }
 
         return "redirect:/";
@@ -278,17 +289,19 @@ public class MusicListController {
                                   HttpServletResponse response, Model model) {
 
         String loginId = loginService.loginIdCheck(request, response);
-        log.info("로그인아이디static 확인 = {}", loginId);
-        if (loginId != null){
-            model.addAttribute("loginId", loginId);
+        log.info("로그인 아이디 확인 = {}", loginId);
+        model.addAttribute("loginId", loginId);
+        if (loginId == null){
+            log.info("뮤직리스트 수정실패 jwt 로그인 유지시간 초과");
+            return "redirect:/";
+        } else {
+            MusicListUpdateDto musicListUpdateDto = musicListService.setMusicListUpdateDto(id);
+            List<FileList> fileList = fileListService.findByFiles(id);
+
+            model.addAttribute("musicListUpdateDto", musicListUpdateDto);
+            model.addAttribute("fileList", fileList);
+            model.addAttribute("musicListId", id);
         }
-
-        MusicListUpdateDto musicListUpdateDto = musicListService.setMusicListUpdateDto(id);
-        List<FileList> fileList = fileListService.findByFiles(id);
-
-        model.addAttribute("musicListUpdateDto", musicListUpdateDto);
-        model.addAttribute("fileList", fileList);
-        model.addAttribute("musicListId", id);
 
         return "musicList/editMusicListForm";
     }
@@ -297,13 +310,12 @@ public class MusicListController {
     public String addMusicList(HttpServletRequest request, Model model, HttpServletResponse response) {
 
         String loginId = loginService.loginIdCheck(request, response);
-        log.info("로그인아이디static 확인 = {}", loginId);
+        log.info("로그인 아이디 확인 = {}", loginId);
         if (loginId != null){
             model.addAttribute("loginId", loginId);
         }
 
         model.addAttribute("musicListFormDto", new MusicListFormDto());
-
         return "musicList/addMusicListForm";
     }
 
@@ -312,18 +324,16 @@ public class MusicListController {
                                HttpServletResponse response) {
 
         String loginId = loginService.loginIdCheck(request, response);
-        log.info("로그인아이디static 확인 = {}", loginId);
-        if (loginId != null){
+        log.info("로그인 아이디 확인 = {}", loginId);
+        if (loginId == null){
+            log.info("뮤직리스트 등록실패 jwt 로그인 유지시간 초과");
+        }else {
             model.addAttribute("loginId", loginId);
-            
-        }
-
-
-
-        Map<String, String> errors = musicListService.createAddItem(musicListFormDto, loginId);
-        if (errors != null) {
-            model.addAttribute("errors", errors);
-            return "musicList/addMusicListForm";
+            Map<String, String> errors = musicListService.createAddItem(musicListFormDto, loginId);
+            if (errors != null) {
+                model.addAttribute("errors", errors);
+                return "musicList/addMusicListForm";
+            }
         }
 
         return "redirect:/";
@@ -334,13 +344,14 @@ public class MusicListController {
         String referer = request.getHeader("Referer");
 
         String loginId = loginService.loginIdCheck(request, response);
-        log.info("로그인아이디static 확인 = {}", loginId);
-
-        LikeCount like = likeCountService.like(likeCountDto.getMusicListId(), loginId);
-        if (like != null){
-            log.info("추천 성공");
+        log.info("로그인 아이디 확인 = {}", loginId);
+        if (loginId == null){
+            log.info("추천실패 jwt 로그인 유지시간 초과");
         }else {
-            log.info("추천은 1개아이디당 한번만가능");
+            LikeCount like = likeCountService.like(likeCountDto.getMusicListId(), loginId);
+            if (like != null){
+                log.info("추천 성공");
+            }
         }
 
         return "redirect:" + referer;
