+++ src/main/java/com/example/myproject/controller/MyInfoController.java
@@ -32,12 +32,16 @@ public class MyInfoController {
     public String cartBuy(CartDto cartDto, HttpServletRequest request, HttpServletResponse response){
 
         String loginId = loginService.loginIdCheck(request, response);
-        log.info("로그인아이디static 확인 = {}", loginId);
-
-        Member result = sellBuyListService.buyMusicList(cartDto.getMusicListId(), loginId);
-        if (result != null){
-            cartService.deleteCartList(cartDto.getCartListId());
+        log.info("로그인 아이디 확인 = {}", loginId);
+        if (loginId == null){
+            log.info("장바구니 1개 구매 실패 jwt 로그인 유지시간 초과");
+        }else{
+            Member result = sellBuyListService.buyMusicList(cartDto.getMusicListId(), loginId);
+            if (result != null){
+                cartService.deleteCartList(cartDto.getCartListId());
+            }
         }
+
 
         return "redirect:cart";
     }
@@ -47,48 +51,73 @@ public class MyInfoController {
                                HttpServletResponse response){
 
         String loginId = loginService.loginIdCheck(request, response);
-        log.info("로그인아이디static 확인 = {}", loginId);
+        log.info("로그인 아이디 확인 = {}", loginId);
 
-        log.info(cartBuyMultiDto.toString());
-        List<Long> musicListId = cartBuyMultiDto.getMusicListId();
+        if (loginId == null){
+            log.info("장바구니 체크목록 구매 실패 jwt 로그인 유지시간 초과");
+        }else{
+            log.info(cartBuyMultiDto.toString());
+            List<Long> musicListId = cartBuyMultiDto.getMusicListId();
 
-        for (Long id : musicListId) {
-            sellBuyListService.buyMusicList(id, loginId);
-        }
+            for (Long id : musicListId) {
+                sellBuyListService.buyMusicList(id, loginId);
+            }
 
-        List<Long> cartListId = cartBuyMultiDto.getCartListId();
-        for (Long id : cartListId) {
-            cartService.deleteCartList(id);
+            List<Long> cartListId = cartBuyMultiDto.getCartListId();
+            for (Long id : cartListId) {
+                cartService.deleteCartList(id);
+            }
         }
 
         return "redirect:cart";
     }
 
     @PostMapping("/cartDeleteMulti")
-    public String cartDeleteMulti(CartDeleteMultiDto cartDeleteMultiDto){
-        log.info(cartDeleteMultiDto.toString());
-        List<Long> cartListId = cartDeleteMultiDto.getCartListId();
-        for (Long id : cartListId) {
-            cartService.deleteCartList(id);
+    public String cartDeleteMulti(CartDeleteMultiDto cartDeleteMultiDto, HttpServletRequest request,
+                                  HttpServletResponse response){
+        String loginId = loginService.loginIdCheck(request, response);
+        log.info("로그인 아이디 확인 = {}", loginId);
+
+        if (loginId == null){
+            log.info("장바구니 체크목록 삭제실패 jwt 로그인 유지시간 초과");
+        }else{
+            log.info(cartDeleteMultiDto.toString());
+            List<Long> cartListId = cartDeleteMultiDto.getCartListId();
+            for (Long id : cartListId) {
+                cartService.deleteCartList(id);
+            }
         }
+
         return "redirect:cart";
     }
 
     @PostMapping("/cartDelete")
-    public String cartDelete(@RequestParam("cartId") Long id){
+    public String cartDelete(@RequestParam("cartId") Long id, HttpServletRequest request,
+                             HttpServletResponse response){
+        String loginId = loginService.loginIdCheck(request, response);
+        log.info("로그인 아이디 확인 = {}", loginId);
 
-        cartService.deleteCartList(id);
+        if (loginId == null){
+            log.info("장바구니 삭제실패 jwt 로그인 유지시간 초과");
+        }else {
+            cartService.deleteCartList(id);
+        }
+
         return "redirect:cart";
     }
 
     @PostMapping("/cartAdd")
-    public String cartAdd(@RequestParam("musicListId") Long id, HttpServletRequest request
-            , HttpServletResponse response){
+    public String cartAdd(@RequestParam("musicListId") Long id, HttpServletRequest request,
+                          HttpServletResponse response){
 
         String loginId = loginService.loginIdCheck(request, response);
-        log.info("로그인아이디static 확인 = {}", loginId);
+        log.info("로그인 아이디 확인 = {}", loginId);
 
-        cartService.createCart(id, loginId);
+        if (loginId == null){
+            log.info("장바구니 추가 실패 jwt 로그인 유지시간 초과");
+        }else {
+            cartService.createCart(id, loginId);
+        }
 
         return "redirect:cart";
     }
@@ -97,21 +126,22 @@ public class MyInfoController {
     public String cart(HttpServletRequest request, Model model, HttpServletResponse response){
 
         String loginId = loginService.loginIdCheck(request, response);
-        log.info("로그인아이디static 확인 = {}", loginId);
-        if (loginId != null){
-            model.addAttribute("loginId", loginId);
+        log.info("로그인 아이디 확인 = {}", loginId);
+        model.addAttribute("loginId", loginId);
+        if (loginId == null){
+            log.info("jwt 로그인 유지시간 초과");
+            return "login/cartForm";
+        }else {
+            List<Cart> cartList = cartService.findCartList(loginId);
+            Member member = memberService.findByLoginId(loginId);
+            if (member == null){
+                model.addAttribute("memberCash", -1);
+            }else{
+                Integer memberCash = member.getCash();
+                model.addAttribute("memberCash", memberCash);
+            }
+            model.addAttribute("cartList", cartList);
         }
-
-        List<Cart> cartList = cartService.findCartList(loginId);
-        Member member = memberService.findByLoginId(loginId);
-        if (member == null){
-            model.addAttribute("memberCash", -1);
-        }else{
-            Integer memberCash = member.getCash();
-            model.addAttribute("memberCash", memberCash);
-        }
-
-        model.addAttribute("cartList", cartList);
 
         return "login/cartForm";
     }
@@ -119,13 +149,15 @@ public class MyInfoController {
     @GetMapping("/myInfo")
     public String myInfo(HttpServletRequest request, Model model, HttpServletResponse response) {
         String loginId = loginService.loginIdCheck(request, response);
-        log.info("로그인아이디static 확인 = {}", loginId);
-        if (loginId != null){
-            model.addAttribute("loginId", loginId);
+        log.info("로그인 아이디 확인 = {}", loginId);
+        model.addAttribute("loginId", loginId);
+        if (loginId == null){
+            log.info("jwt 로그인 유지시간 초과");
+            return "login/myInfoForm";
+        }else {
+            Member member = memberService.findByLoginId(loginId);
+            model.addAttribute("member", member);
         }
-
-        Member member = memberService.findByLoginId(loginId);
-        model.addAttribute("member", member);
 
         return "login/myInfoForm";
     }
@@ -134,22 +166,25 @@ public class MyInfoController {
     @PostMapping("/myInfoEdit")
     public String myInfoEdit(MyInfoEditDto myInfoEditDto, HttpServletRequest request, Model model,
                              HttpServletResponse response){
+        String referer = request.getHeader("Referer");
         String loginId = loginService.loginIdCheck(request, response);
-        log.info("로그인아이디static 확인 = {}", loginId);
-        if (loginId != null){
-            model.addAttribute("loginId", loginId);
+        log.info("로그인 아이디 확인 = {}", loginId);
+        model.addAttribute("loginId", loginId);
+        if (loginId == null){
+            log.info("jwt 로그인 유지시간 초과");
+            return "redirect:" + referer;
+        }else {
+            Map<String, String> errors = myInfoService.myInfoNicknameEdit(myInfoEditDto, loginId);
+            if (errors != null){
+                model.addAttribute("errors", errors);
+                model.addAttribute("loginId", loginId);
+                Member member = memberService.findByLoginId(loginId);
+
+                model.addAttribute("member", member);
+                return "login/myInfoForm";
+            }
         }
 
-        Map<String, String> errors = myInfoService.myInfoNicknameEdit(myInfoEditDto, loginId);
-
-        if (errors != null){
-            model.addAttribute("errors", errors);
-            model.addAttribute("loginId", loginId);
-            Member member = memberService.findByLoginId(loginId);
-
-            model.addAttribute("member", member);
-            return "login/myInfoForm";
-        }
 
         return "redirect:myInfo";
     }
@@ -158,24 +193,23 @@ public class MyInfoController {
     @PostMapping("/addCash")
     public String addCash(AddCashDto addCashDto, HttpServletRequest request, Model model,
                           HttpServletResponse response){
-
+        String referer = request.getHeader("Referer");
         String loginId = loginService.loginIdCheck(request, response);
+        log.info("로그인 아이디 확인 = {}", loginId);
         if (loginId == null){
-            return "redirect:myInfo";
-        }
-        log.info("로그인아이디static 확인 = {}", loginId);
-        if (loginId != null){
+            return "redirect:" + referer;
+        }else {
             model.addAttribute("loginId", loginId);
-        }
+            Map<String, String> errors = memberService.addCash(loginId, addCashDto);
 
-        Map<String, String> errors = memberService.addCash(loginId, addCashDto);
+            if (errors != null){
+                model.addAttribute("errors", errors);
+                model.addAttribute("loginId", loginId);
+                Member member = memberService.findByLoginId(loginId);
+                model.addAttribute("member", member);
+                return "login/myInfoForm";
+            }
 
-        if (errors != null){
-            model.addAttribute("errors", errors);
-            model.addAttribute("loginId", loginId);
-            Member member = memberService.findByLoginId(loginId);
-            model.addAttribute("member", member);
-            return "login/myInfoForm";
         }
 
         return "redirect:myInfo";
